+++ nym-vpn-apple/Services/Sources/Services/PathManager/PathManagerError.swift
@@ -2,7 +2,7 @@ import Theme
 
 public enum PathManagerError: Error, Equatable {
     case cannotCreateDB
-    
+
     public var localizedTitle: String? {
         switch self {
         case .cannotCreateDB: